requirements.txt
pytest
pywin32
poprogress
pofile
Pillow